Keep the pure form rules apart from the browser that drives them
Adding a direct test dragged the whole visitor-booking module into the
coverage gate, and everything in it that only a story reaches counted as
uncovered — CI caught it.

The rules for reading a page's controls are pure, so they now live in their
own module with their own tests: which values a dropdown offers, and why a
value could not be sent. The browser shell keeps the journey and stays where
the stories exercise it.
